+++ assets/javascripts/kitten/components/typography/sup-title/index.js
@@ -1,30 +1,17 @@
 import React from 'react'
-import classNames from 'classnames'
-import styled from 'styled-components'
-import { stepToRem } from '../../../helpers/utils/typography'
-import TYPOGRAPHY from '../../../constants/typography-config'
+import { Text } from '../../../components/typography/text'
 
-const StyledSupTitle = styled.div`
-  font-size: ${stepToRem(-1)};
-  line-height: normal;
-  text-transform: uppercase;
-  ${TYPOGRAPHY.fontStyles.bold};
-`
-
-export const SupTitle = ({
-  className,
-  tag,
-  ...supTitleProps
-}) => (
-
-  <StyledSupTitle
-    as={tag}
-    className={classNames('k-SupTitle', className)}
-    {...supTitleProps}
+export const SupTitle = ({ className, ...textProps}) => (
+  <Text
+    size="tiny"
+    lineHeigth="normal"
+    transform="uppercase"
+    weight="bold"
+    className={className}
+    {...textProps}
   />
 )
 
 SupTitle.defaultProps = {
-  tag: 'div',
   children: 'Lorem ipsum',
 }